ARM: dts: AM4372: add dma properties for uart
Add DMA properties for uart modules instances where
direct mapped channels are available. For other uarts
which are on EDMA crossbar, the mapping needs to be done
based on channel usage on the board they are used on.

AM437x UARTs are compatible to AM335x UARTs in terms of
DMA quirks required, so add "ti,am3352-uart" to the
compatible list. This does not break backward compatibility
as existing DTBs should continue to work with newer kernels
albeit without DMA support.

Signed-off-by: Sekhar Nori <nsekhar@ti.com>