ARM: dts: dra7: workaround UART module disable errata
DRA7x UARTs are compatible to AM335x UARTs in terms of
DMA quirk required.

Due to Advisory 1.0.41 as documented in AM335x errata document,
UART module cannot be disabled once DMA is used. The only
workaround is to softreset the module before disabling it.

Add "ti,am3352-uart" to the compatible list so the driver
workaround for this errata is enabled.

This does not break backward compatibility as existing DTBs
should continue to work with newer kernels albeit without DMA
support.

Signed-off-by: Sekhar Nori <nsekhar@ti.com>