HACK: drivers/rtc/rtc-ds1307.c: Ensure OUT bit is set to prevent infinite interrupts on MCP7941x
Based on Data sheet for MCP79410:
http://ww1.microchip.com/downloads/en/DeviceDoc/20002266F.pdf

When CONTROL::SQWEN is 0 and when CONTROL::ALM0EN is 0, the signal
polarity depends on CONTROL::OUT bit

The polarity of interrupt depends on CONTROL:ALM0EN to be generated and
ALM0WKDAY::ALMPOL bit.

Currently driver sets ALM0WKDAY::ALMPOL as 0, however does not control
OUT - this maybe high or low, which means that on systems which depend
on level interrupts based on ALMPOL, disabling ALM0EN will end up
generating infinite interrupts.

A proper fix would be to configure the OUT and ALMPOL based on IRQ type
requested for the platform. For the moment, hack things to be compliant
with configuration.

Signed-off-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>