HACK: ARM: dts: am57xx_beagle_x15: Fix IRQ type for mcp7941x
Interrupt signal from MCP79410 on the current platform looks like the following:
	+---------+  +--+             +------->
		  |  |  |             |
		  |  |  |             |
		  |  |  |             |
		  +--+  +-------------+
		    ^^

Using rising edge, we detect interrupt on the glitch which happens
to be generated, and works as a result. However, there are lots of
unclear questions to be answered.

As per GIC distributor TRM:
http://infocenter.arm.com/help/topic/com.arm.doc.ddi0438d/BGBHIACJ.html#BABJFCFB

ARM GIC distributor does not support IRQ trigger type
IRQ_TYPE_LEVEL_LOW, then why did it even worked in the first
place(when gicd was never configured) - aka Why does GIC respond
even though GIC configuration for LOW cannot be supported? When,
due to some other issue that is yet to be determined, when external
source provides default level as low, why does GIC generate infinite
interrupt when GICD can only support rising edge or

Why does GIC generate an interrupt even if we change the mux away from
SYS_NIRQ to something else?

There are lots of unanswered questions that need to be addressed for .
this to be finalized HACK the situation for the moment by switching  .
to use IRQ_TYPE_EDGE_RISING                                          .

Signed-off-by: Grygorii Strashko <Grygorii.Strashko@linaro.org>
Signed-off-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>