ARM: OMAP2+: cpuidle33xx: Change wkup_m3 idle state to avoid MPU PLL bypass
To get to C1-state in cpuidle am33xx currently uses the wkup_m3 to gate
the MPU clock domain and bypass the MPU PLL. Because we are not shutting
off the MPU power domain in this C-state it is possible for both to be
awake and executing at the same time, and we do not have the ability
to synchronize execution between the two in the cpuidle path due to
no interrupt context.

This leads to racy behavior between the two and in certain instances
where c1-state is entered often during periods of high cpu activity it
is possible for the MPU to interfere with the wake path on the wkup_m3,
which prevents the MPU PLL from being re-locked and causing extreme
system slowdown.

Because of this, we must now use a different idle state available on the
current wkup_m3 firmware (0x190) which does nothing more than put the
MPU clockdomain to sleep, giving us the same power savings seen
previously on the MPU power rail but slightly higher power on the
MPU_PLL voltage rail.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>