ARM: dts: DRA7: Remove obsolete 7 inch display
7 inch is no longer plan of record. 10 inch display is POR. no point
in maintaining additional code That needs constant update w.r.t
changes as we discover it.

As part of this change, get rid of mxt244 which is present only on 7
inch display.

In fact, if one of the 7 inch displays go bad, we cannot even get a
replacement display from the vendor manufacturing the 7 inch display,
not to mention the additional overhead of needing to maintain the
custom touchscreen controller driver(which does not work with device
tree in upstream without significant support burden) etc.

Signed-off-by: Nishanth Menon <nm@ti.com>
Acked-by: Tomi Valkeinen <tomi.valkeinen@ti.com>
Signed-off-by: Jyri Sarha <jsarha@ti.com>