remoteproc/pruss: fix incorrect failure reporting of pru_rproc_kick
The mbox_send_message() API returns a zero or a positive token id on
success, and a negative value on failure. The pru_rproc_kick() function
is incorrectly checking for a non-zero return value to report a failure,
even though the mailbox message has been transmitted successfully, so
fix the error condition check properly.

Signed-off-by: Jason Reeder <jreeder@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>