drivers/rtc/rtc-ds1307.c: disable the right irq handler in handler
The irq handler is split into a work which is scheduled by the irq
handler. while wakeirq can only trigger once, the same handler handles
both the regular ISR and the wakeirq.

However the wakeirq handling is slightly different as it controlled from
above layers and is "pseudo interrupt". we must ensure that the correct
irq is disabled in irq handler and re-enabled in client handler, else,
we end up with unbalanced irq enable/disable in the system.

Fixes: 05bed213b91f ("drivers/rtc/rtc-ds1307.c: Support optional wakeup interrupt source")
Reported-by: Grygorii Strashko <Grygorii.Strashko@linaro.org>
Signed-off-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>