ARM: dts: am437x-gp-evm: Add regulator-suspend-enable for RTC DCDCs
TPS65218 present on am437x-gp-evm has defined that DCDC5 and DCDC6 must
be removed from sequencer control in order for them to remain active
when PMIC_PWR_EN line is pulled low during a poweroff or RTC-mode
transition. Add regulator-suspend-enable flag to the corresponding
nodes to do this.

On A1 revision of the TPS65218, FSEAL bit would be undefined without
coin-cell present which in many cases led to it being set, causing DCDC5
and DCDC6 to stay active and allowing RTC-only mode to work but also
leading to unexplained failures when it was not. On B1 revision, FSEAL
is always 0 when no coin-cell is present so this patch is required on
boards with B1 revision to ever work. This implementation works on
boards with either A1 or B1 revision and makes sure that DCDC5 and DCDC6
always stay active.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>