mmc: omap_hsmmc: Fix card enumeration failure when card plug/unplug multiple times
When sd card is removed during an ongoing data transfer, the card
enumeration fails on next insertion.
Usually, when data transfer request completes, the TC/CC interrupt
handler and dma completion callback does the cleanup of the mmc_request
structure. The TC/CC interrupt is not raised if the card is removed in
the middle of data transfer.  Therefore, the mmc core does not give up
the controller over host controller thinking that there is still an
ongoing data transfer.  Hence on the next insertion the mmc_rescan()
enters uninterruptible sleep waiting for previous release to happen.

It is possible to do cleanup during card removal event but the bottom
half of TC/CC interrupt or the dma completion callback might execute
after card removal interrupt in which case there will be an oops.
So the fix is to clean up any pending mmc_request during the insertion
event before passing on the event to mmc core.

Signed-off-by: Vignesh R <vigneshr@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>