drm/omap: fix memory leak when omap_plane_mode_set fails
When the user updates a plane with a new framebuffer, but with illegal
settings, we leave the framebuffer unreleased, leaking memory.

This is because omap_plane_update() refs the fb, but doesn't unref it in
case of failure.

It would be nice to handle this the same way as in mainline
f2d022aa421ca903a30f63b04528064b7eceaf5e ("drm/omap: fix the handling of
fb ref counts"), but without primary planes that cannot be done.

So this patch tweaks omap_plane_update() to handle the fb ref management
correctly.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>
Signed-off-by: Jyri Sarha <jsarha@ti.com>