mailbox: Add ability for clients to request channels by name
[ Upstream commit dfabde206aa10ae71a89ba75e68b1f58a6336a05 ]

This patch supplies a new framework API; mbox_request_channel_byname().

It works by supplying the usual client pointer as the first argument and
a string as the second.  The API will search the client's node for a
'mbox-names' property then request a channel in the normal way using the
requested string's index as the expected second 'index' argument.

Signed-off-by: Lee Jones <lee.jones@linaro.org>
Signed-off-by: Jassi Brar <jaswinder.singh@linaro.org>
[s-anna@ti.com: cherry-pick commit 'dfabde206aa1' from v4.2]
Signed-off-by: Suman Anna <s-anna@ti.com>