OMAPDSS: sii9022: fix audio locking
sii9022 driver's audio code does not take the mutex used on the video
side. Normally this doesn't cause any problems, but when the video side
is reading EDID, it must switch the SiI9022 to a DDC mode, which messes
up the normal i2c transactions. This can lead to the following error:

[ 288.344464] sii9022 1-003b: ASoC: can't set hdmi-hifi hw params: -121

This patch adds the locking to audio side.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>
Signed-off-by: Jyri Sarha <jsarha@ti.com>