OMAPDSS: sii9022: retry edid read
When reading EDID, usually at boot time with built-it drivers, we
sometimes get -EREMOTEIO, failing the EDID read.

The reason is unclear, but doing an i2c read to SiI9022 after requesting
the DDC access seems to cause -EREMOTEIO. We don't do that, but it could
mean that SiI9022 has some issues around the DDC access.

Looks like a simple retry of the EDID read solves the issue, or at least
makes the problem happen very rarely.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>
Signed-off-by: Jyri Sarha <jsarha@ti.com>