phy: ti-pipe3: fix SATA when AHCI_PLATFORM is m and SATA enabled in u-boot
When SATA is used by u-boot and not powered down correctly *and*
if AHCI_PLATFORM is m, clock auto-disable will kick in
and disable sata_refclk thus causing Errata i783 and breaking SATA.

Since we can't guarantee what's done in u-boot, we take a safe approach by
keeping the sata_refclk ON in pipe3 probe so that it doesn't get auto-disabled
by the kernel. We keep track of this with the sata_refclk_enabled flag
and do an additional disable in ti_pipe3_disable_refclk() when needed.

Signed-off-by: Roger Quadros <rogerq@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>