iommu/omap: add support for runtime auto suspend/resume
This patches adds the support for the OMAP IOMMUs to be suspended
during the auto suspend/resume of the OMAP remoteproc devices. The
remote processors are auto suspended after a certain time of idle
or inactivity period. This is done by extending the suspend/resume
API added with an additional flag to indicate the auto-suspend.
The runtime auto-suspend simply decrements and increments the
runtime usage count of the IOMMU devices and let the context be
saved/restored using the existing runtime pm callbacks.

Signed-off-by: Suman Anna <s-anna@ti.com>