iommu/omap: introduce new API for suspend/resume
The MMU registers for the remote processors lose their context
in Open Switch Retention (OSWR) or device OFF modes. Hence, the
context of the IOMMU needs to be saved before it is put to these
lower power state (OSWR/OFF) and restored before it is powered
up to ON again. The IOMMUs need to be active as long as the
client devices that are present behind the IOMMU are active.

This patch introduces two new API, omap_iommu_domain_suspend()
and omap_iommu_domain_resume() to allow the client users of the
IOMMU devices to suspend & resume the IOMMU devices during their
suspend/resume operations. The API leverages the pm runtime
framework API to implement suspend/resume functionality, and
invoke the runtime PM callbacks. The PM runtime_suspend and
runtime_callbacks already are used to enable, configure and
disable the IOMMUs during the attaching and detaching of the
client devices to the IOMMUs, and this code is reused during
suspend/resume.

NOTE:
There are two other existing API, omap_iommu_save_ctx() and
omap_iommu_restore_ctx(). These are left as is to support
suspend/resume of devices on legacy OMAP3 SoC.

Signed-off-by: Suman Anna <s-anna@ti.com>