extcon: usb-gpio: handle EPROBE_DEFER for gpio requests
The extcon us gpio driver requests id and vbus gpios during
probe. This could fail with -EPROBE_DEFER in case the gpiochip
that has the specific gpio pin is not initialised. For eg: the
gpiochip could be an i2c client and is not probed at the time that
the extcon driver gets probed.

Propagating the EPROBE_DEFER error lets the extcon driver probe
function to be called again until it succeeds or error out with
something other than EPROBE_DEFER.

Acked-by: Roger Quadros <rogerq@ti.com>
Signed-off-by: Subramaniam Chanderashekarapuram <subramaniam.ca@ti.com>
[nsekhar@ti.com: move EPROBE_DEFER check inside IS_ERR() check]
Signed-off-by: Sekhar Nori <nsekhar@ti.com>