media: ti-vpe: vip: Different IRQ naming for each slice IRQ
VIP device has two interrupts and the driver uses one for each slice.
It's easier to differentiate IRQs from each of the slice if they are
named differently.

Use the V4L2 device name for naming the interrupts so that interrupts
from different slices can be differentiated when accessed from procfs.

Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>
Acked-by: Benoit Parrot <bparrot@ti.com>
Signed-off-by: Jyri Sarha <jsarha@ti.com>