media: i2c: ov1063x: Fix OV1063x supported YUV 422 formats
The sensor was only listing YUYV as the only supported format but
it was setting the sensor to generate UYVY instead.
This inversion was set incorrectly because the bridge driver
is grabbing the bus data out of order. This needs to be fixed in the
bridge driver instead. So the sensor should advertise and generate
consistent pixel format.

In addition it was found that the sensor was able to support all 4
variants of YUV 422 interleaved formats.

So the driver will know correctly advertised and support:
UYVY, YUYV, VYUY and YVYU.

Signed-off-by: Benoit Parrot <bparrot@ti.com>
Signed-off-by: Jyri Sarha <jsarha@ti.com>