ARM: OMAP2+: pm33xx: Fix RTC wakeup src handling
Once we enable EXT_WAKEUP in the RTC_PMIC register all wakeup events
report as EXT_WAKEUP because the RTC_WAKEUP line on the SOC is connected
to the nWAKEUP signal of the PMIC which generates an event on any PMIC
power on event, whether it is a pushbutton wake or PMIC_PWR_EN change
from RTC alarm event.

Previously we check the RTC_PMIC register to determine the wakeup source
but now we can just check the RTC_STATUS register to see if RTC_ALARM1
status bit is set, and if it is report RTC_ALARM as wake source,
otherwise report another EXT_WAKEUP.

This is important not only for reporting the correct wake source but
also making sure that the RTC IRQ is set for retrigger_irq in
am33xx_pm_end, otherwise the ALARM1 status does not get cleared and if
an RTC mode transition is made with out using rtcwake (with the
intention of using PB to wake) the board will immediately wake due to
ALARM1 status and fail to enter RTC mode.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>