ti_config_fragments/connectivity.cfg: disable CONFIG_USB_OTG_WHITELIST
CONFIG_USB_OTG_WHITELIST is to be enabled for compliance to OTG
specification. However, on TI SoCs, we only support/test USB
dual-role device, not the full OTG specification.

Moreover, with CONFIG_USB_OTG_WHITELIST enabled, on AM335x
beaglebone black, on connecting a USB webcam we get:

[   74.960459] musb-hdrc musb-hdrc.1.auto: otg: usb_otg_kick_fsm: invalid host/gadget device
[   75.080228] usb 1-1: new high-speed USB device number 4 using musb-hdrc
[   75.270151] usb 1-1: device v03f0 pa707 is not supported
[   75.275670] hub 1-0:1.0: unable to enumerate USB device on port 1

This is because kernel refuses to enumerate peripherals which
are not listed in kernel's OTG whitelist. This is needed for
full compliance with OTG specification.

Disable CONFIG_USB_OTG_WHITELIST since we support only dual-role
device and do not need to reject non-whitelisted devices during
enumeration.

Suggested-by: Roger Quadros <rogerq@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>