rtc: omap: setup the regulators for poweroff mode
In certain cases, like rtc-only mode, the system regulators shall be
set to slightly different mode compared to suspend to mem. As the RTC
driver is handling the poweroff setup on some systems, add a callback
to regulator core from the RTC power off to setup the valid regulator
states.

Signed-off-by: Tero Kristo <t-kristo@ti.com>
Signed-off-by: Dave Gerlach <d-gerlach@ti.com>