OMAPDSS: dra7-tpd12s015: move ddc pixmux to .dts
We need to do runtime pinmuxing for HDMI, as the HDMI DDC lines are also
used as I2C2 lines. This is currently done in the DRA7 EVM specific
tpd12s015 driver, by writing directly to the pinmux registers.

Remove that hackery by moving the muxing to .dts.

We also need to use a McASP2 pin, which is used as a gpio via McASP2
functionality (i.e. not in GPIO mode). Muxing for this pin is also moved
to .dts, but changing the output value of the pin is still handled in
the driver via direct register writes.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>
Tested-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Jyri Sarha <jsarha@ti.com>