OMAPDSS: dra7-tpd12s015: support system suspend
The driver uses mcasp2 pin for a gpio, with direct register writes. On
system suspend the mcasp2 is left alive, preventing suspend.

Also, mcasp2 may be used for other purposes, so directly touching the
mcasp2 registers can cause problems. Luckily, the same pin can be muxed
to mcasp8 mode, and mcasp8 is not used by anyone.

We now have hacks made to mcasp driver which allows the dra7-tpd12s015
driver to manage the mcasp8 pin via the mcasp driver. While the solution
is a pure hack, it allows us to do proper power management, and remove
all the direct register write hackery.

dra7-tpd12s015's device tree data contains a phandle to the mcasp8 node,
and dra7-tpd12s015 driver uses this handle to enable and disable the
mcasp8 module, and to set the gpio output value.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>
Tested-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Jyri Sarha <jsarha@ti.com>