remoteproc/omap: revise the comment about standby in _suspend
The _suspend() function has a TODO comment about implementing a
BIOS-independent hook to send a message to notify the MPU the
completion of the context save on the remoteproc side. This has
been clarified with the SYS/BIOS team now, and this race condition
between the remoteproc saving the context and the MPU waiting for
the sync point before it puts the remote processor into reset, can
never be resolved using a message ACK or a shared memory variable
signalling. So, update the comment appropriately.

Signed-off-by: Suman Anna <s-anna@ti.com>